backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-genai
requests
edge-tts
Pillow
tenacity
sqlalchemy
pymysql
passlib
bcrypt
email-validator
langfuse
